minor updates and bug fixes in analysis_logger.py
- fixed bug in auto specifying watchfolder
- added flag in analysis_config.json to suppress autocreating dailynb

+++ analysis_logger.py
@@ -192,7 +192,6 @@ def main(analysis_type, watchfolder, load_matlab=True, images_per_shot=1, save_i
                             id=str(popped_id[0]), file=os.path.join(watchfolder, 'run_ids.txt')))
 
                 print('\n')
-
         time.sleep(refresh_time)
 
 
@@ -204,9 +203,11 @@ if __name__ == '__main__':
     try:
         analysis_paths = load_analysis_path()
         data_basepath = analysis_paths['data_basepath']
+        notebook_flag = (analysis_paths['create_notebook_flag'] == 'True')
     except FileNotFoundError:
         data_basepath = ''
         print('analysis_config.json not found, using default settings.')
+        notebook_flag = True
     bec1server_path = load_bec1serverpath()
     analysis_shorthand = {'zt': 'zcam_triple_imaging',
                           'zd': 'zcam_dual_imaging',
@@ -226,7 +227,8 @@ if __name__ == '__main__':
                 print(name)
                 last_output = name
     watchfolder = measurement_directory(
-        measurement_name=suggest_run_name(newrun_input='n', appendrun_input='y', basepath=data_basepath))
+        measurement_name=suggest_run_name(newrun_input='n', appendrun_input='y', basepath=data_basepath),
+        basepath = data_basepath)
     save_images = True
     save_images_input = input('Keep images after analysis? [y/n]: '
                               )
@@ -234,12 +236,12 @@ if __name__ == '__main__':
         print('entering testing mode...')
         save_images = False
 
-    clean_notebook_path = r'D:\Fermidata1\enrico\log viewer and plotterDUPLICATEANDUSETODAY.ipynb'
+    clean_notebook_path = os.path.join(os.path.dirname(__file__), 'log viewer and plotterDUPLICATEANDUSETODAY.ipynb')
     if analysis_key == 'y':
         clean_notebook_path = r'C:\Users\FermiCam2\Desktop\GitHub\enrico\log viewer and plotterDUPLICATEANDUSETODAY.ipynb'
     nb_path = os.path.join(os.path.dirname(watchfolder), 'dailynb.ipynb')
 
-    if not os.path.exists(nb_path):
+    if not os.path.exists(nb_path) and notebook_flag:
         shutil.copy(clean_notebook_path, nb_path)
         print('no daily notebook, made a duplicate from {path} now.'.format(
             path=clean_notebook_path))
